backend/requirements.txt
websockets==12.0
openai==1.0.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
